sessao15/requirements.txt
crewai>=0.28.0
langchain>=0.0.335
langchain-community>=0.0.1
openai>=1.3.7
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
langchainhub>=0.1.13
